- Fixed bug #125: show_mem_delta does not calculate correct negative values on   64bit machines.
SVN Rev: 1835

+++ xdebug_private.h
@@ -151,8 +151,8 @@ typedef struct _function_stack_entry {
 	HashTable   *symbol_table;
 
 	/* tracing properties */
-	unsigned int memory;
-	unsigned int prev_memory;
+	signed long  memory;
+	signed long  prev_memory;
 	double       time;
 
 	/* profiling properties */
